# swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "App Store Connect Adapters",
    "description": "HTTP API for querying an App Store Connect adapters host.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:48189"
    }
  ],
  "paths": {
...
    "/api/app-store-connect/v1.0/tags/{adapterId}/properties": {
      "post": {
        "tags": [
          "TagSearch"
        ],
        "summary": "Gets tag property definitions from an adapter.",
        "operationId": "TagSearch_GetTagPropertiesAll",
        "parameters": [
          {
            "name": "adapterId",
            "in": "path",
            "required": true,
            "description": "The adapter ID.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "The search filter.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetTagPropertiesRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "Successful responses contain a collection of AdapterProperty objects.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdapterProperty"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "TagSearch"
        ],
        "summary": "Gets tag property definitions from an adapter.",
        "operationId": "TagSearch_GetTagProperties",
        "parameters": [
          {
            "name": "adapterId",
            "in": "path",
            "required": true,
            "description": "The adapter ID.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "The page size.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "x-position": 2
          },
          {
            "name": "page",
            "in": "query",
            "description": "The results page to return.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "Successful responses contain a collection of AdapterProperty objects.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagDefinition"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/app-store-connect/v1.0/tags/{adapterId}/find": {
      "post": {
        "tags": [
          "TagSearch"
        ],
        "summary": "Performs a tag search on an adapter.",
        "operationId": "TagSearch_FindTagsAll",
        "parameters": [
          {
            "name": "adapterId",
            "in": "path",
            "required": true,
            "description": "The adapter ID.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "The search filter.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FindTagsRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "Successful responses contain a collection of TagDefinition objects.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagDefinition"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "TagSearch"
        ],
        "summary": "Performs a tag search on an adapter.",
        "operationId": "TagSearch_FindTags",
        "parameters": [
          {
            "name": "adapterId",
            "in": "path",
            "required": true,
            "description": "The adapter ID.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "name",
            "in": "query",
            "description": "The tag name filter.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "description",
            "in": "query",
            "description": "The tag description filter.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "units",
            "in": "query",
            "description": "The tag units filter.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 4
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "The page size.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            },
            "x-position": 5
          },
          {
            "name": "page",
            "in": "query",
            "description": "The results page to return.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            },
            "x-position": 6
          }
        ],
        "responses": {
          "200": {
            "description": "Successful responses contain a collection of TagDefinition objects.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagDefinition"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/app-store-connect/v1.0/tags/{adapterId}/get-by-id": {
      "post": {
        "tags": [
          "TagSearch"
        ],
        "summary": "Gets tags by ID.",
        "operationId": "TagSearch_GetTagsAll",
        "parameters": [
          {
            "name": "adapterId",
            "in": "path",
            "required": true,
            "description": "The adapter ID.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "request",
          "description": "The request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetTagsRequest"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "Successful responses contain a collection of matching TagDefinition objects.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagDefinition"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "TagSearch"
        ],
        "summary": "Gets tags by ID.",
        "operationId": "TagSearch_GetTags",
        "parameters": [
          {
            "name": "adapterId",
            "in": "path",
            "required": true,
            "description": "The adapter ID.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "tag",
            "in": "query",
            "style": "form",
            "explode": true,
            "description": "The IDs of the tags to retrieve.",
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "type": "string"
              }
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "Successful responses contain a collection of matching TagDefinition objects.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TagDefinition"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PageableAdapterRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AdapterRequest"
          },
          {
            "type": "object",
            "description": "Base class for adapter requests that support paging.",
            "x-abstract": true,
            "additionalProperties": false,
            "properties": {
              "pageSize": {
                "type": "integer",
                "description": "The page size for the query.",
                "format": "int32",
                "default": 10,
                "maximum": 2147483647.0,
                "minimum": 1.0
              },
              "page": {
                "type": "integer",
                "description": "The page number for the query.",
                "format": "int32",
                "default": 1,
                "maximum": 2147483647.0,
                "minimum": 1.0
              }
            }
          }
        ]
      },
      "AdapterRequest": {
        "type": "object",
        "description": "Base class that adapter request objects can inherit from.",
        "x-abstract": true,
        "additionalProperties": false,
        "properties": {
          "properties": {
            "type": "object",
            "description": "Additional request properties. These can be used to provide bespoke query parameters \nsupported by the adapter.",
            "nullable": true,
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
...
      "AdapterProperty": {
        "type": "object",
        "description": "Describes a custom property associated with an adapter, tag, tag value, event message, etc.",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The property name.",
            "minLength": 1
          },
          "value": {
            "description": "The property value.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/Variant"
              }
            ]
          },
          "description": {
            "type": "string",
            "description": "The property description.",
            "nullable": true
          }
        }
      },
      "Variant": {
        "type": "object",
        "description": "Describes a variant value.",
        "additionalProperties": false,
        "properties": {
          "value": {
            "description": "The value.",
            "nullable": true
          },
          "type": {
            "description": "The variant type.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/VariantType"
              }
            ]
          }
        }
      },
      "VariantType": {
        "type": "string",
        "description": "Describes the type of a variant value.",
        "x-enumNames": [
          "Unknown",
          "Null",
          "Object",
          "Boolean",
          "SByte",
          "Byte",
          "Int16",
          "UInt16",
          "Int32",
          "UInt32",
          "Int64",
          "UInt64",
          "Float",
          "Double",
          "String",
          "DateTime",
          "TimeSpan",
          "Url"
        ],
        "enum": [
          "Unknown",
          "Null",
          "Object",
          "Boolean",
          "SByte",
          "Byte",
          "Int16",
          "UInt16",
          "Int32",
          "UInt32",
          "Int64",
          "UInt64",
          "Float",
          "Double",
          "String",
          "DateTime",
          "TimeSpan",
          "Url"
        ]
      },
...
      "TagIdentifier": {
        "type": "object",
        "description": "Defines basic information for identifying a real-time data tag.",
        "additionalProperties": false,
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique identifier for the tag.",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "description": "The tag name.",
            "minLength": 1
          }
        }
      },
...
      "GetTagPropertiesRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PageableAdapterRequest"
          },
          {
            "type": "object",
            "description": "Describes a request to retrieve tag property definitions.",
            "additionalProperties": false
          }
        ]
      },
      "TagDefinition": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TagSummary"
          },
          {
            "type": "object",
            "description": "Describes a tag definition.",
            "additionalProperties": false,
            "properties": {
              "states": {
                "type": "array",
                "description": "The discrete states for the tag. If DataType is not \nInt32, this property will be null.",
                "items": {
                  "$ref": "#/components/schemas/DigitalState"
                }
              },
              "properties": {
                "type": "array",
                "description": "Bespoke tag properties.",
                "items": {
                  "$ref": "#/components/schemas/AdapterProperty"
                }
              },
              "labels": {
                "type": "array",
                "description": "Labels associated with the tag.",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
      "DigitalState": {
        "type": "object",
        "description": "Describes a digital state associated with a tag.",
        "additionalProperties": false,
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The state name.",
            "minLength": 1
          },
          "value": {
            "type": "integer",
            "description": "The state value.",
            "format": "int32"
          }
        }
      },
      "TagSummary": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TagIdentifier"
          },
          {
            "type": "object",
            "description": "Describes summary information about a tag.",
            "additionalProperties": false,
            "properties": {
              "description": {
                "type": "string",
                "description": "The tag description.",
                "nullable": true
              },
              "units": {
                "type": "string",
                "description": "The tag units.",
                "nullable": true
              },
              "dataType": {
                "description": "The tag data type.",
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/VariantType"
                  }
                ]
              }
            }
          }
        ]
      },
      "FindTagsRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PageableAdapterRequest"
          },
          {
            "type": "object",
            "description": "Describes a tag search query.",
            "additionalProperties": false,
            "properties": {
              "name": {
                "type": "string",
                "description": "The tag name filter.",
                "nullable": true
              },
              "description": {
                "type": "string",
                "description": "The tag description filter.",
                "nullable": true
              },
              "units": {
                "type": "string",
                "description": "The tag units filter.",
                "nullable": true
              },
              "label": {
                "type": "string",
                "description": "The tag label filter.",
                "nullable": true
              },
              "other": {
                "type": "object",
                "description": "Additional filters on bespoke tag properties.",
                "nullable": true,
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
      "GetTagsRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AdapterRequest"
          },
          {
            "type": "object",
            "description": "Describes a request to get tag definitions by tag ID or name.",
            "additionalProperties": false,
            "required": [
              "tags"
            ],
            "properties": {
              "tags": {
                "type": "array",
                "description": "The IDs or names of the tags to retrieve.",
                "minItems": 1,
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
...
    }
  }
}